Rename module advanced_search to search_filters
